Use ItemMeta in catalog lookups and confine test doubles to test_support.
