+++ crates/sqyre-executor/src/backends.rs
@@ -1,16 +1,15 @@
-//! Domain-coupled executor ports and test doubles.
+//! Domain-coupled executor ports.
 //!
 //! OS-facing traits live in [`sqyre_ports`] and are re-exported here so executor
 //! consumers keep a single import path.
 
-use image::RgbaImage;
 use sqyre_domain::{CoordinateRef, Macro};
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
 pub use sqyre_ports::{
-    clamp_search_rect, AutomationBackend, AutomationError, CaptureError, DesktopRect, MoveOptions,
-    PortError, RgbCapture, ScreenCapturer, WindowFocuser,
+    clamp_search_rect, AutomationBackend, AutomationError, CaptureError, DesktopRect, ItemMeta,
+    MoveOptions, PortError, RgbCapture, ScreenCapturer, WindowFocuser,
 };
 
 /// Resolve `program~point` / search-area refs using the loaded program catalog.
@@ -42,14 +41,6 @@ pub trait IconStore {
     /// Optional mask PNG for the item (resized by caller).
     fn mask_path(&self, target: &str) -> Option<std::path::PathBuf>;
     fn item_meta(&self, target: &str) -> Option<ItemMeta>;
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct ItemMeta {
-    pub name: String,
-    pub stack_max: i32,
-    pub cols: i32,
-    pub rows: i32,
 }
 
 /// Look up another macro by name.
@@ -105,234 +96,3 @@ impl OcrEngine for sqyre_vision::LeptessOcr {
     }
 }
 
-/// Test OCR engine that returns a fixed result.
-#[derive(Debug, Default)]
-pub struct FixedOcrEngine {
-    pub result: sqyre_vision::OcrRecognition,
-    pub log: std::sync::Mutex<Vec<String>>,
-}
-
-impl OcrEngine for FixedOcrEngine {
-    fn recognize(
-        &self,
-        image: &sqyre_match::ImageBuf,
-    ) -> Result<sqyre_vision::OcrRecognition, PortError> {
-        if let Ok(mut g) = self.log.lock() {
-            g.push(format!(
-                "ocr:{}x{}c{}",
-                image.width, image.height, image.channels
-            ));
-        }
-        Ok(self.result.clone())
-    }
-}
-
-/// Test OCR engine that pops results from a FIFO queue (then repeats the last).
-#[derive(Debug, Default)]
-pub struct QueuedOcrEngine {
-    pub queue: std::sync::Mutex<Vec<sqyre_vision::OcrRecognition>>,
-    pub log: std::sync::Mutex<Vec<String>>,
-}
-
-impl OcrEngine for QueuedOcrEngine {
-    fn recognize(
-        &self,
-        image: &sqyre_match::ImageBuf,
-    ) -> Result<sqyre_vision::OcrRecognition, PortError> {
-        if let Ok(mut g) = self.log.lock() {
-            g.push(format!(
-                "ocr:{}x{}c{}",
-                image.width, image.height, image.channels
-            ));
-        }
-        let mut q = self
-            .queue
-            .lock()
-            .map_err(|_| PortError::Message("QueuedOcrEngine: lock poisoned".into()))?;
-        if q.is_empty() {
-            return Err(PortError::invalid("QueuedOcrEngine: empty queue"));
-        }
-        if q.len() == 1 {
-            return Ok(q[0].clone());
-        }
-        Ok(q.remove(0))
-    }
-}
-
-/// Recording backend for unit tests.
-#[derive(Debug, Default)]
-pub struct RecordingBackend {
-    pub log: Vec<String>,
-}
-
-impl AutomationBackend for RecordingBackend {
-    fn milli_sleep(&mut self, ms: i32) {
-        self.log.push(format!("sleep:{ms}"));
-    }
-    fn move_to(&mut self, x: i32, y: i32, opts: MoveOptions) {
-        self.log
-            .push(format!("move:{x},{y},smooth={}", opts.smooth));
-    }
-    fn click(&mut self, button: &str, down: bool) -> Result<(), AutomationError> {
-        self.log.push(format!(
-            "click:{button}:{}",
-            if down { "down" } else { "up" }
-        ));
-        Ok(())
-    }
-    fn scroll(&mut self, up: bool) -> Result<(), AutomationError> {
-        self.log
-            .push(format!("scroll:{}", if up { "up" } else { "down" }));
-        Ok(())
-    }
-    fn key_down(&mut self, key: &str) -> Result<(), AutomationError> {
-        self.log.push(format!("keydown:{key}"));
-        Ok(())
-    }
-    fn key_up(&mut self, key: &str) -> Result<(), AutomationError> {
-        self.log.push(format!("keyup:{key}"));
-        Ok(())
-    }
-    fn type_char(&mut self, ch: char) {
-        self.log.push(format!("type:{ch}"));
-    }
-    fn write_clipboard(&mut self, s: &str) -> Result<(), AutomationError> {
-        self.log.push(format!("clipboard:{s}"));
-        Ok(())
-    }
-}
-
-/// In-memory capturer for tests.
-#[derive(Debug, Default)]
-pub struct RecordingCapturer {
-    pub log: Vec<String>,
-    /// Single image returned when [`Self::queue`] is empty.
-    pub next: Option<RgbaImage>,
-    /// FIFO images consumed one per capture (then falls back to [`Self::next`]).
-    pub queue: Vec<RgbaImage>,
-    pub bounds: DesktopRect,
-}
-
-impl RecordingCapturer {
-    fn take_image(&mut self) -> Result<RgbaImage, crate::CaptureError> {
-        if !self.queue.is_empty() {
-            return Ok(self.queue.remove(0));
-        }
-        self.next
-            .clone()
-            .ok_or_else(|| crate::CaptureError::Message("RecordingCapturer: no image".into()))
-    }
-}
-
-impl ScreenCapturer for RecordingCapturer {
-    fn capture_monitor(&mut self, display_index: i32) -> Result<RgbaImage, crate::CaptureError> {
-        self.log.push(format!("monitor:{display_index}"));
-        self.take_image()
-    }
-    fn capture_rect(&mut self, rect: DesktopRect) -> Result<RgbaImage, crate::CaptureError> {
-        self.log
-            .push(format!("rect:{},{},{},{}", rect.x, rect.y, rect.w, rect.h));
-        self.take_image()
-    }
-    fn virtual_bounds(&mut self) -> Result<DesktopRect, crate::CaptureError> {
-        Ok(self.bounds)
-    }
-}
-
-/// In-memory macro catalog for tests.
-#[derive(Debug, Default)]
-pub struct MapMacroLookup {
-    pub macros: std::collections::BTreeMap<String, Arc<Macro>>,
-}
-
-impl MacroLookup for MapMacroLookup {
-    fn get(&self, name: &str) -> Option<Arc<Macro>> {
-        self.macros.get(name).cloned()
-    }
-}
-
-/// Test waiter that returns immediately (does not block).
-#[derive(Debug, Default)]
-pub struct ImmediateContinueWaiter {
-    pub log: std::sync::Mutex<Vec<String>>,
-    /// Indices returned by successive `wait_for_any_chord` calls (defaults to 0).
-    pub any_queue: std::sync::Mutex<Vec<usize>>,
-}
-
-impl ContinueKeyWaiter for ImmediateContinueWaiter {
-    fn wait_for_continue(
-        &self,
-        keys: &[String],
-        pass_through: bool,
-        _stop: &AtomicBool,
-    ) -> Result<(), PortError> {
-        if keys.is_empty() {
-            return Err(PortError::invalid("pause: continue key not set"));
-        }
-        if let Ok(mut g) = self.log.lock() {
-            g.push(format!(
-                "continue:{}:passthrough={pass_through}",
-                keys.join("+")
-            ));
-        }
-        Ok(())
-    }
-
-    fn wait_for_any_chord(
-        &self,
-        chords: &[Vec<String>],
-        hold_repeat: &[bool],
-        pass_through: bool,
-        _stop: &AtomicBool,
-    ) -> Result<usize, PortError> {
-        if chords.is_empty() || chords.iter().all(|c| c.is_empty()) {
-            return Err(PortError::invalid("key wait: no chords configured"));
-        }
-        let idx = self
-            .any_queue
-            .lock()
-            .ok()
-            .and_then(|mut q| {
-                if q.is_empty() {
-                    None
-                } else {
-                    Some(q.remove(0))
-                }
-            })
-            .unwrap_or(0);
-        if let Ok(mut g) = self.log.lock() {
-            let labels: Vec<String> = chords
-                .iter()
-                .enumerate()
-                .map(|(i, c)| {
-                    let hold = if hold_repeat.get(i).copied().unwrap_or(false) {
-                        "*"
-                    } else {
-                        ""
-                    };
-                    format!("{hold}{}", c.join("+"))
-                })
-                .collect();
-            g.push(format!(
-                "any:{}:pick={idx}:passthrough={pass_through}",
-                labels.join("|")
-            ));
-        }
-        Ok(idx.min(chords.len().saturating_sub(1)))
-    }
-}
-
-/// Test focuser that records calls.
-#[derive(Debug, Default)]
-pub struct RecordingWindowFocuser {
-    pub log: std::sync::Mutex<Vec<String>>,
-}
-
-impl WindowFocuser for RecordingWindowFocuser {
-    fn focus(&self, process_path: &str, window_title: &str) -> Result<(), AutomationError> {
-        if let Ok(mut g) = self.log.lock() {
-            g.push(format!("focus:{process_path}:{window_title}"));
-        }
-        Ok(())
-    }
-}
