+++ crates/sqyre-executor/src/run.rs
@@ -704,7 +704,8 @@ pub(crate) fn resolve_text(text: &str, macro_: &Macro) -> Result<String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::backends::{DesktopRect, RecordingBackend, RecordingCapturer};
+    use crate::backends::DesktopRect;
+    use crate::test_support::{RecordingBackend, RecordingCapturer};
     use crate::test_support::FixedResolver;
     use sqyre_domain::{
         root_loop, Action, ActionId, ActionKind, CoordinateRef, ScalarValue, VariableAssignment,
